src/backend/pyproject.toml
[tool.poetry]
name = "rfid-kube"
version = "0.1.0"
description = ""
authors = ["Roland Sobczak <rolandsobczak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.112.0"}
sqlalchemy = "^2.0.32"
kubernetes = "^30.1.0"
psycopg2-binary = "^2.9.9"
sqlalchemy-utils = "^0.41.2"
pydantic = {extras = ["email"], version = "^2.8.2"}
pika = "^1.3.2"
gunicorn = "^23.0.0"
environs = "^11.0.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
